+++ src/layout/LayoutApp.jsx
@@ -3,6 +3,7 @@ import { Outlet, ScrollRestoration } from "react-router-dom";
 import LeafFall from "../components/LeafFall";
 import Footer from "./components/Footer";
 import Header from "./components/Header";
+import Disclaimer from "../components/disclaimer/Disclaimer";
 
 function LayoutApp() {
   return (
@@ -10,6 +11,7 @@ function LayoutApp() {
       <LeafFall />
       <div className="relative z-10 flex h-full w-full flex-col items-center justify-between">
         <Header />
+        <Disclaimer />
         <ScrollRestoration />
         <Outlet />
         <Footer />